requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.1
bandit==1.6.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
check-requirements-txt==1.0.2
cloudinary==1.24.0
colorama==0.4.3
cryptography==3.4.7
cssselect==1.1.0
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.8
django-aggregate-if==0.5
django-allauth==0.44.0
django-cors-headers==3.7.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-filter==2.4.0
django-location-field==2.1.0
djangorestframework==3.11.1
dodgy==0.2.1
fake-useragent==0.1.11
filelock==3.0.12
flake8==3.8.3
flake8-polyfill==1.0.2
geographiclib==1.50
geopy==2.0.0
gitdb==4.0.5
GitPython==3.1.7
gunicorn==20.0.4
heroku3==4.2.3
httpie==2.2.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
mypy==0.782
mypy-extensions==0.4.3
oauthlib==3.1.0
parse==1.18.0
pbr==5.4.5
pep8-naming==0.10.0
Pillow==7.2.0
pipenv==2020.8.13
prospector==1.3.0
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.0
pyee==7.0.4
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==2.1.0
pylama==7.7.1
pylint==2.5.2
pylint-celery==0.3
pylint-django==2.0.15
pylint-flask==0.6
pylint-plugin-utils==0.6
pyppeteer==0.2.2
pyquery==1.4.3
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.0
requirements-detector==0.7
setoptconf==0.2.0
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
soupsieve==2.0.1
sqlparse==0.3.1
stdlib-list==0.8.0
stevedore==3.2.0
toml==0.10.1
tqdm==4.54.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.10
virtualenv==20.0.28
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
w3lib==1.22.0
websockets==8.1
whitenoise==5.2.0
wrapt==1.12.1
